LayoutTests/media/modern-media-controls/tracks-support/mac/tracks-support-show-panel-fullscreen.html

<!DOCTYPE html>
<script src="../../../../resources/js-test-pre.js"></script>
<script src="../../../../resources/ui-helper.js"></script>
<script src="../../resources/media-controls-utils.js" type="text/javascript"></script>
<body>
<video src="../../../content/CC+Subtitles.mov" style="width: 320px; height: 240px;" controls autoplay></video>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Showing the tracks panel in fullscreen.");

setTimeout(() => {
    debug("FAIL: did not end fast enough");
    finishJSTest();
}, 10000);

internals.setMockVideoPresentationModeEnabled(true);

const media = document.querySelector("video");

media.addEventListener("webkitfullscreenchange", () => {
    debug("webkitfullscreenchange event is fired");
    shouldBecomeEqual("internals.isChangingPresentationMode(media)", "false", () => {
        if (!media.webkitDisplayingFullscreen) {
            media.remove();
            finishJSTest();
        }

        shouldBecomeDifferent(`internals.shadowRoot(media).querySelector("button.tracks")`, "null", () => {
            shouldBecomeDifferent(`internals.shadowRoot(media).querySelector("button.tracks").getBoundingClientRect().width`, "0", () => {
                shouldBecomeEqual(`internals.shadowRoot(media).querySelector("button.tracks").getBoundingClientRect().right`, "582", async () => {
                    await UIHelper.renderingUpdate();
                    pressOnElement(internals.shadowRoot(media).querySelector("button.tracks"));
                    shouldBecomeDifferent(`internals.shadowRoot(media).querySelector('.tracks-panel')`, "null", () => {
                        shouldBeEqualToString("internals.shadowRoot(media).querySelector('.tracks-panel').style.right", "218px");
                        shouldBeEqualToString("internals.shadowRoot(media).querySelector('.tracks-panel').style.bottom", "101px");
                        debug("");
                        media.webkitExitFullscreen();
                    });
                });
            });
        });
    });
});

media.addEventListener("play", () => {
    media.pause();
    shouldBecomeDifferent(`internals.shadowRoot(media).querySelector("button.fullscreen")`, "null", () => {
        shouldBecomeDifferent(`internals.shadowRoot(media).querySelector("button.fullscreen").getBoundingClientRect().width`, "0", () => {
            pressOnElement(internals.shadowRoot(media).querySelector(`button.fullscreen`));
        });
    });
});

</script>
<script src="../../../../resources/js-test-post.js"></script>
</body>
